fix ft_max & ft_min to return size_t

+++ includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: qle-guen <qle-guen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/09/25 18:28:07 by qle-guen          #+#    #+#             */
-/*   Updated: 2016/04/05 23:42:45 by qle-guen         ###   ########.fr       */
+/*   Updated: 2016/04/05 23:52:46 by qle-guen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -82,8 +82,8 @@ int					ft_isprint(int c);
 char				*ft_itoa(int n);
 int					ft_atoi(const char *s);
 int					ft_pow(int a, int n);
-int					ft_max(int a, int b);
-int					ft_min(int a, int b);
+size_t				ft_max(size_t a, size_t b);
+size_t				ft_min(size_t a, size_t b);
 int					ft_isupper(int c);
 int					ft_islower(int c);
 void				ft_putintlst(t_list *t);
